Support for wrapped secrets
Adds a wrap_ttl option to vault_generic_secret so wrapped secrets
can be requested.
It also adds a command option so write operations can also be performed,
this allows to obtain secrets that are requested with write operations
as secret IDs in AppRole auth or keys in PKI module.