token needs to be able to create child tokens
The Vault token passed to Terraform needs to have a policy associated with it that gives it the capability to create child tokens. Specifically, it needs the update capability on the Vault path auth/token/create.  See https://www.vaultproject.io/api/auth/token/index.html#create-token.  A Vault policy that would give the given token the ability to do that is:

```
path "auth/token/create" {
   capabilities = ["update"]
}
```