Fix V2GenericSecret test value.
The test case for updating a v2 generic secret had a check function that
hardcoded in the initial value, and then proceeded to check that against
the updated value, which would never work.

This updates to add a check function for the updated config, with the
updated value coded in.