Always set the backend path in state with no trailing /.
0.12 complained about our plan changing because we were interpolating a
backend's path. Backend paths cannot be specified using a trailing / but
are always set in state with a trailing /. That means the plan will
always change when creating a backend and using it in the same request.

This commit changes that to set the backend in state with no trailing
slash, which resolves the issue because it will always match user input
now. This may be considered a breaking change.