Fix double list.
The basic LDAP auth backend group test had policies specified as a list,
with the util function for turning a slice into a Terraform list inside
that list. Except the util helper also returned the square brackets, so
we had `[[item, item]]` instead of `[item, item]`. In Terraform 0.11,
this was fine. In Terraform 0.12, it is not.