Make DB statements a list for Postgres backend.
Make our postgres database secret backend use a list of statements for
creation, revocation, rollback, and renew, as the API has updated to use
a list for those fields. This is a breaking change, as we're changing
the schema type, but we coded in support for previous versions of the
API that returned those fields as strings, as well.