Fix invalid d.Sets.
The bound principal arns for the AWS auth role backend were being set
incorrectly, because we were keying off what the user had in the config
to determine whether the response should be a string or a list. But the
server always responded with a list, no matter what was sent. We updated
to remove the string version, as we're doing a major release anyways,
and to test the response type before setting it so we don't set the
wrong type.

Login was also trying to set a lease_id that doesn't exist in the
schema, so we just removed that Set call.