Don't remove cassandra values from state.
Don't remove the values from the database_secret_backend_connection
resource's cassandra block that the API doesn't return. Because we
overwrite the whole block in state, we need to fall back on what's in
state, which presumably came from the config.

We'll miss any drift, but the API doesn't surface that information, so
there's not much we can do about that.