ldap_auth_backend_user: fix diff when groups is empty (#655)
Vault stores `groups` for an LDAP user as a string (though it stores
`policies` for an LDAP user as a list). The `ldap_auth_backend_user`
uses `schema.TypeSet` for the `groups` key - correctly so, as it
provides a cleaner API for consumers. However, this type mismatch
resulted in a permanent diff when some set of groups was set to an
LDAP user, then subsequently removed.

This changes adds an explicit check for when the `groups` key that
is read is an empty string. If that is the case, then we want to
simply use an empty set for diffing purposes and short-circuit any
string splitting or parsing.