Add region passing for non commercial regions (#841)
* Add support for passing region information to vault backend

This allows us to use non-inferrable regions

* Remove unnecessary data passing when obtaining region

* Remove root privilege check, add region as part of data source

This allows the tf config to specify the region. It creates redundancy,
however, it should allow people to use a different region without special
permissions of any sort. If no region is given, it should default to
us-east-1.

* Rely on go zero values for modifying aws.Config

This allows us to set the region ONLY if it was specified in the data section of
the user's Terraform config.