TODO: vault: Removed the use of 'helper/encryption' package
As documented on https://www.terraform.io/docs/extend/guides/v2-upgrade-guide.html#removal-of-helper-encryption-package is
not recomended and not recommended to encrypt parts of the TFState.

For now I did just remove the import so I can proceed with the import but his change is major and it may even need to change
the schema and remove some values like the 'encrypted_client_token' as it'll not be needed anymore.

OFC we can still have this logic using another lib/code that encrypts, but for now I just removed the use to continue with
the migration