Added example of AWS auth method with header_value
Currently, none of the parameters mentioned in the Vault API docs specify how to pass in the `X-Vault-AWS-IAM-Server-ID` as part of the provider block.  There is a mention of `iam_server_id_header_value` in the API doc, but that key does not work in the provider's parameters.  

We identified that `header_value` is what the provider's parameter block needs and validated with a customer that it works.