Rename optional child token config item to skip_child_token
Per pull request review from @benashz
This name change includes a flip in the logic: if skipping the
creation of the child token is true, we do not create it and
assign it, if false we follow the default behavior of using
the child token. Tests and documentation were updated to match.