Disable logging HTTP request/response body (#1251)
Logging either the request or response body exposes secrets when
Terraform is executed with a log level of DEBUG or higher. For Vault,
this exposes secrets in clear text which should be avoided. However It
is still possible to log request/response body via the following
environment variables:

* `TERRAFORM_VAULT_LOG_BODY` - when set to `true` both the request and
  response body will be logged.

* `TERRAFORM_VAULT_LOG_REQUEST_BODY` - when set to `true` the request
body will be logged.

* `TERRAFORM_VAULT_LOG_RESPONSE_BODY` - when set to `true` the
response body will be logged.

Co-authored-by: Tom Proctor <tomhjp@users.noreply.github.com>