🐛 Arg removal not properly applied for aws_secret_backend_role
When an existing argument like `policy_document` is removed for
`aws_secret_backend_role`, the change won't
be properly applied to Vault, as it's only checking
for existing data: If there is a removal, Terraform will simply
not send the data to Vault. This results in constant
attempts of Terraform to remove the data
without success.

This changes the checks for what data should be sent to Vault,
so arguments will always be set in case their data has changed
(even if it's a change to null).