Handle the case where member_identity_ids may be nil
During a plan operation the identity_group_member_entity_ids resource
reads the target identity group from Vault. If the identity group
exists, and the resource is configured in non-exclusive mode, the
resource fails to check the result of the type assertion. This results
in an interface conversion panic.

Fix is to check the type assertion before further processing of the
identity group's member_identity_ids field.