Rework the credential validation to use azidentity
Credential validation now relies on listing of the resource principles
to determine success. If the client credentials can successfully list
the principles once then the credentials provided from Vault are deemed
to be valid. In the case where the 401 HTTP status code is returned from
the Azure API, the validation check is terminated since this means that
the credentials were not authorized to perform any API requests.