Fix issue where group member entities were lost (#1409)
The identity_group resource was removing the member_entity_ids when
configured with external_member_entity_ids. In this case the
member_entity_ids are handled externally and should not be modified by
the identity_group resource. The issue only manifested on resource
update when any of the tracked fields had changes.

Fixes:
- completely re-work the tests for
  identity_group_member_entity_ids, making them more reliable
- add test cases that include changing a identity_group's group_name,
  which triggers the issue in the identity_group resource
- make set type operations simpler
- numerous other fixes
- add extra group name change to the dynamic test