Fix race condition on duplicate entity alias creation (#1429)
The creation of entity aliases should be serialized, since we want to
handle the case when the user is attempting to create duplicate entity
aliases during the same Terraform execution. A duplicate entity alias
shares the same mount accessor and name with another entity alias.

The fix is to lock the /identity/entity-alias/id path on all create,
update, and delete operations. This allows the provider to reliably
detect the duplicate entity alias condition, and prevent duplicate
alias creation.

Other fixes:
- all CRUD functions are of the *Context type, and can return meaning
  diag.Diagnostics
- created a new entity package under internal/identity/entity with some
  helper functions