Properly force a new resource whenenver the cert is near expiry
Affected resources:
- resource_pki_secret_backend_cert
- resource_pki_secret_backend_sign

The previous approach would leave some computed values out of sync
whenever the certificate was within the expiry window. This was
because it attempted to leverage a combination of schema CustomizeDiff,
and Update functions, rather than taking advantage of the ability to
mark the resource as requiring replacement with the CustomizeDiff
function itself.