Support setting namespace per resource (#1305)
This change adds support for configuring namespaces at the resource or data source level.

All namespace directives are applied relative to the provider's configured namespace. This new approach allows for namespace'd resources to be created without having to pass a namespace specific provider, although that method is still fully supported.