Add data source for fetching the Vault address from the provider
Fetch the URL of the root of the target Vault server. Allows getting the
URL configured for the provider.

The data source can be useful if the address is configured for the
provider with the `VAULT_ADDR` environment variable, or if the URL is
needed in a child module using the provider.