Fix GCP auth with service account credentials
We are trying to use the provider with SA credentials, but the
authentication does not succeed because the required scope is not
present in the JWT request:
"https://www.googleapis.com/auth/cloud-platform"

This scope is mandatory according to the documentation https://developer.hashicorp.com/vault/docs/auth/gcp#gcp-credentials
The JWT sign request fails with the following error:
```
{"error":"invalid_scope","error_description":"Invalid OAuth scope or ID token audience provided."}
```

I was able to reproduce this in the TestAuthLoginGCP_Login test, but
only after setting the initialized:true setting on the
AuthLoginGCP.AuthLoginCommon object. I guess the test is not ran when
this value is not positioned.

This fixes the JWT token generation, but the provider's call to Vault is
rejected because the `credentials` and `service_account` parameters are
not expected by the endpoint. I added them to the `copyParamsExcluding`
call and changed the expected parameters in the test.

I can now run this test with valid credentials:

```
$ GOOGLE_APPLICATION_CREDENTIALS="/tmp/creds.json"
TF_ACC_GOOGLE_SERVICE_ACCOUNT="xxx@yyyy.iam.gserviceaccount.com" go test -run TestAuthLoginGCP_Login
PASS
ok      github.com/hashicorp/terraform-provider-vault/internal/provider 0.491s
```