Ensure that the qr_size can be properly configured.  (#1750)
Previously, the identity_mfa_totp resource's schema did not allow the
qr_size field to be configured by the user. It was set as Computed,
but not Optional. After resolving this issue we discovered that
Create/Update functions would always include this param in the vault API
request, which negated the benefits of the field being Computed.

In order to resolve the above issue above, and other similar ones, it is
now possible to set the VaultAPIGetterFunc by field on the
contextFuncConfig. Which means it is possible to use any of the
following schema.ResourceData methods: GetOk(), Get(), GetOkExists()

* Drop redundant getGetter()

* Ensure that qr_size can be set to a 0 value