Support authenticating to the root namespace from within an auth_login* (#2066)
* Add new auth_login field for enforcing root namespace

The additional field is a work around to deal with TF plugin SDK's
evaluation of values in boolean context. There was no way to treat the
"" as the default Vault namespace from the TF schema, since none of its
Get*() public methods do what we need. So the new way is to set
is_root_namespace to true, leave the namespace field unset in any
auth_login* resource blocks.

---------

Co-authored-by: Vinay Gopalan <vinay@hashicorp.com>