feat(azure): add explicit_max_ttl for azure roles
Add support for configuring `explicit_max_ttl` for Azure Secrets roles.
This field was released as part of Azure Secrets plugin v0.20.0,
incorporated into Vault 1.18.0 and later.

Note that the field is configured as an integer to align with how other
resources throughout the provider handle TTL fields. Configuring it to
a seconds-based integer avoids Terraform flagging drifts between the
seconds returned from Vault's API and the Terraform string-based
configuration.